feature(cart-persistence): changing cart write stream to no longer cache

+++ src/Ambev.DeveloperEvaluation.Application/Carts/DeleteCart/DeleteCartHandler.cs
@@ -15,7 +15,12 @@ namespace Ambev.DeveloperEvaluation.Application.Carts.DeleteCart
 
         public async Task<DeleteCartResult> Handle(DeleteCartCommand command, CancellationToken cancellationToken)
         {
-            // Adicione validação se necessário
+            var validator = new DeleteCartCommandValidator();
+            var validationResult = await validator.ValidateAsync(command, cancellationToken);
+
+            if (!validationResult.IsValid)
+                throw new ValidationException(validationResult.Errors);
+
             var success = await _cartRepository.DeleteAsync(command.Id, cancellationToken);
             return new DeleteCartResult { Success = success };
         }
